requirements.txt
python-dotenv==1.0.1
openai==1.52.2
chromadb==0.5.15
PyPDF2==3.0.1
python-docx
pyyaml
questionary

